Foundation
I have finished the home page which should later be reworked as an introduction page because it will not be a good page to be used as the root of all pages.
I made the lay out, added a button to change pages, added a second page, added an image, formatted the words, images and buttons, and lastly I installed a few missing dependencies.