Fixed time and removed useless fetch weather
The issue with the time was that it wasn't displaying weekday on both operating systems. toLocaleString on android printed too much where toLocaleDateString printed too little. It always worked on IOS, but I decided to just remove the weekday from IOS and add a function that determines what day of the week it is and add it to the string. I essentially manually did what the function should have automatically did on Android.
Also I removed the fetchweather because due to: time constraints, an improper app set up, and many issues in the attempt to re-render information. All of these kept us from needing a way to update weather with a new location.